Moved playground projects to .NET Core
