+++ Source/Samples/Yamo.Playground.CS/Model/Article.cs
@@ -4,11 +4,15 @@ using System.Linq;
 using System.Text;
 using System.Threading.Tasks;
 
-namespace Yamo.PlaygroundCS.Model
+namespace Yamo.Playground.CS.Model
 {
-    class Category
+    class Article
     {
         public int Id { get; set; }
+        public decimal Price { get; set; }
+
         public Label Label { get; set; }
+        public List<ArticlePart> Parts { get; set; }
+        public List<Category> Categories { get; set; }
     }
 }
