+++ Source/Samples/Yamo.Playground.CS/Model/ArticlePart.cs
@@ -4,15 +4,14 @@ using System.Linq;
 using System.Text;
 using System.Threading.Tasks;
 
-namespace Yamo.PlaygroundCS.Model
+namespace Yamo.Playground.CS.Model
 {
-    class Article
+    class ArticlePart
     {
         public int Id { get; set; }
+        public int ArticleId { get; set; }
         public decimal Price { get; set; }
 
         public Label Label { get; set; }
-        public List<ArticlePart> Parts { get; set; }
-        public List<Category> Categories { get; set; }
     }
 }
