+++ Source/Samples/Yamo.Playground.CS/Model/Category.cs
@@ -4,14 +4,11 @@ using System.Linq;
 using System.Text;
 using System.Threading.Tasks;
 
-namespace Yamo.PlaygroundCS.Model
+namespace Yamo.Playground.CS.Model
 {
-    class ArticlePart
+    class Category
     {
         public int Id { get; set; }
-        public int ArticleId { get; set; }
-        public decimal Price { get; set; }
-
         public Label Label { get; set; }
     }
 }
