+++ Source/Samples/Yamo.Playground.CS/Model/PropertyModifiedTrackingBase.cs
@@ -4,7 +4,7 @@ using System.Linq;
 using System.Text;
 using System.Threading.Tasks;
 
-namespace Yamo.PlaygroundCS.Model
+namespace Yamo.Playground.CS.Model
 {
     public abstract class PropertyModifiedTrackingBase : IHasDbPropertyModifiedTracking
     {
