+++ Source/Samples/Yamo.Playground.CS/Yamo.Playground.CS.csproj
@@ -1,0 +1,35 @@
+<Project Sdk="Microsoft.NET.Sdk">
+
+  <PropertyGroup>
+    <OutputType>Exe</OutputType>
+    <TargetFramework>netcoreapp3.1</TargetFramework>
+    <OptionStrict>On</OptionStrict>
+    <Version>0.6.0</Version>
+    <Authors>Peter Šimno</Authors>
+    <Company>Peter Šimno</Company>
+    <Description>Yamo playground (C#)</Description>
+    <PackageLicenseExpression>MIT</PackageLicenseExpression>
+    <PackageProjectUrl>https://github.com/esentio/Yamo</PackageProjectUrl>
+    <PackageTags>orm;micro-orm;sql;database</PackageTags>
+    <AssemblyName>Yamo.Benchmarks</AssemblyName>
+    <DisableImplicitNamespaceImports>true</DisableImplicitNamespaceImports>
+  </PropertyGroup>
+
+  <ItemGroup>
+    <Import Include="System" />
+    <Import Include="System.Collections" />
+    <Import Include="System.Collections.Generic" />
+    <Import Include="System.Linq" />
+    <Import Include="System.Threading.Tasks" />
+  </ItemGroup>
+
+  <ItemGroup>
+    <PackageReference Include="Microsoft.Data.SqlClient" Version="2.0.0" />
+  </ItemGroup>
+
+  <ItemGroup>
+    <ProjectReference Include="..\..\Source\Yamo.SqlServer\Yamo.SqlServer.vbproj" />
+    <ProjectReference Include="..\..\Source\Yamo\Yamo.vbproj" />
+  </ItemGroup>
+
+</Project>
