+++ Source/Samples/Yamo.Playground.VB/Model/Article.vb
