+++ Source/Samples/Yamo.Playground.VB/Model/ArticlePart.vb
