+++ Source/Samples/Yamo.Playground.VB/Model/Blog.vb
