+++ Source/Samples/Yamo.Playground.VB/Model/Comment.vb
