+++ Source/Samples/Yamo.Playground.VB/Model/Label.vb
