+++ Source/Samples/Yamo.Playground.VB/Test.vb
