+++ Source/Yamo.sln
@@ -13,10 +13,6 @@ Project("{778DAE3C-4631-46EA-AA77-85C1314464D9}") = "Yamo.SqlServer", "Source\Ya
 EndProject
 Project("{2150E333-8FDC-42A3-9474-1A3956D46DE8}") = "Samples", "Samples", "{3E03BA6A-76DC-4B42-8C54-99FDD716ED70}"
 EndProject
-Project("{F184B08F-C81C-45F6-A57F-5ABD9991F28F}") = "Yamo.Playground", "Samples\Yamo.Playground\Yamo.Playground.vbproj", "{64B7C964-21DA-4612-9317-35E0F5562BE1}"
-EndProject
-Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "Yamo.PlaygroundCS", "Samples\Yamo.PlaygroundCS\Yamo.PlaygroundCS.csproj", "{04EFC854-905D-4875-85EB-0AAEEFFE5F47}"
-EndProject
 Project("{778DAE3C-4631-46EA-AA77-85C1314464D9}") = "Yamo.SQLite", "Source\Yamo.SQLite\Yamo.SQLite.vbproj", "{9CF4DB78-78F3-4E40-A60B-807BE00CD02C}"
 EndProject
 Project("{2150E333-8FDC-42A3-9474-1A3956D46DE8}") = "Benchmarks", "Benchmarks", "{16FA0B29-C6FF-46A8-BB2D-AC9CA744BA07}"
@@ -31,11 +27,15 @@ Project("{778DAE3C-4631-46EA-AA77-85C1314464D9}") = "Yamo.CodeGenerator", "Sourc
 EndProject
 Project("{778DAE3C-4631-46EA-AA77-85C1314464D9}") = "Yamo.Benchmarks", "Benchmarks\Yamo.Benchmarks\Yamo.Benchmarks.vbproj", "{49E61955-B655-4752-ABD1-D7802B69BF44}"
 EndProject
-Project("{F184B08F-C81C-45F6-A57F-5ABD9991F28F}") = "Yamo.Test", "Test\Yamo.Test\Yamo.Test.vbproj", "{F9B10063-5389-4F5E-8F2B-4C2E46988EC6}"
+Project("{778DAE3C-4631-46EA-AA77-85C1314464D9}") = "Yamo.Test", "Test\Yamo.Test\Yamo.Test.vbproj", "{F9B10063-5389-4F5E-8F2B-4C2E46988EC6}"
 EndProject
-Project("{F184B08F-C81C-45F6-A57F-5ABD9991F28F}") = "Yamo.Test.SQLite", "Test\Yamo.Test.SQLite\Yamo.Test.SQLite.vbproj", "{E1DE7318-1459-4AAD-9A36-90DE438BD4C6}"
+Project("{778DAE3C-4631-46EA-AA77-85C1314464D9}") = "Yamo.Test.SQLite", "Test\Yamo.Test.SQLite\Yamo.Test.SQLite.vbproj", "{E1DE7318-1459-4AAD-9A36-90DE438BD4C6}"
 EndProject
-Project("{F184B08F-C81C-45F6-A57F-5ABD9991F28F}") = "Yamo.Test.SqlServer", "Test\Yamo.Test.SqlServer\Yamo.Test.SqlServer.vbproj", "{FEC0BB53-CB6D-48F2-BBF8-6D82AF5B7647}"
+Project("{778DAE3C-4631-46EA-AA77-85C1314464D9}") = "Yamo.Test.SqlServer", "Test\Yamo.Test.SqlServer\Yamo.Test.SqlServer.vbproj", "{FEC0BB53-CB6D-48F2-BBF8-6D82AF5B7647}"
+EndProject
+Project("{F184B08F-C81C-45F6-A57F-5ABD9991F28F}") = "Yamo.Playground.VB", "Samples\Yamo.Playground.VB\Yamo.Playground.VB.vbproj", "{1E64FEBC-B19E-4C94-9F5B-39CB449D0FB7}"
+EndProject
+Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "Yamo.Playground.CS", "Samples\Yamo.Playground.CS\Yamo.Playground.CS.csproj", "{AE5EEFC7-B685-41AF-89D0-C2DDB925C507}"
 EndProject
 Global
 	GlobalSection(SolutionConfigurationPlatforms) = preSolution
@@ -51,14 +51,6 @@ Global
 		{AE939C31-706D-47C2-8445-9A5141AD6A37}.Debug|Any CPU.Build.0 = Debug|Any CPU
 		{AE939C31-706D-47C2-8445-9A5141AD6A37}.Release|Any CPU.ActiveCfg = Release|Any CPU
 		{AE939C31-706D-47C2-8445-9A5141AD6A37}.Release|Any CPU.Build.0 = Release|Any CPU
-		{64B7C964-21DA-4612-9317-35E0F5562BE1}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
-		{64B7C964-21DA-4612-9317-35E0F5562BE1}.Debug|Any CPU.Build.0 = Debug|Any CPU
-		{64B7C964-21DA-4612-9317-35E0F5562BE1}.Release|Any CPU.ActiveCfg = Release|Any CPU
-		{64B7C964-21DA-4612-9317-35E0F5562BE1}.Release|Any CPU.Build.0 = Release|Any CPU
-		{04EFC854-905D-4875-85EB-0AAEEFFE5F47}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
-		{04EFC854-905D-4875-85EB-0AAEEFFE5F47}.Debug|Any CPU.Build.0 = Debug|Any CPU
-		{04EFC854-905D-4875-85EB-0AAEEFFE5F47}.Release|Any CPU.ActiveCfg = Release|Any CPU
-		{04EFC854-905D-4875-85EB-0AAEEFFE5F47}.Release|Any CPU.Build.0 = Release|Any CPU
 		{9CF4DB78-78F3-4E40-A60B-807BE00CD02C}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
 		{9CF4DB78-78F3-4E40-A60B-807BE00CD02C}.Debug|Any CPU.Build.0 = Debug|Any CPU
 		{9CF4DB78-78F3-4E40-A60B-807BE00CD02C}.Release|Any CPU.ActiveCfg = Release|Any CPU
@@ -83,6 +75,14 @@ Global
 		{FEC0BB53-CB6D-48F2-BBF8-6D82AF5B7647}.Debug|Any CPU.Build.0 = Debug|Any CPU
 		{FEC0BB53-CB6D-48F2-BBF8-6D82AF5B7647}.Release|Any CPU.ActiveCfg = Release|Any CPU
 		{FEC0BB53-CB6D-48F2-BBF8-6D82AF5B7647}.Release|Any CPU.Build.0 = Release|Any CPU
+		{1E64FEBC-B19E-4C94-9F5B-39CB449D0FB7}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
+		{1E64FEBC-B19E-4C94-9F5B-39CB449D0FB7}.Debug|Any CPU.Build.0 = Debug|Any CPU
+		{1E64FEBC-B19E-4C94-9F5B-39CB449D0FB7}.Release|Any CPU.ActiveCfg = Release|Any CPU
+		{1E64FEBC-B19E-4C94-9F5B-39CB449D0FB7}.Release|Any CPU.Build.0 = Release|Any CPU
+		{AE5EEFC7-B685-41AF-89D0-C2DDB925C507}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
+		{AE5EEFC7-B685-41AF-89D0-C2DDB925C507}.Debug|Any CPU.Build.0 = Debug|Any CPU
+		{AE5EEFC7-B685-41AF-89D0-C2DDB925C507}.Release|Any CPU.ActiveCfg = Release|Any CPU
+		{AE5EEFC7-B685-41AF-89D0-C2DDB925C507}.Release|Any CPU.Build.0 = Release|Any CPU
 	EndGlobalSection
 	GlobalSection(SolutionProperties) = preSolution
 		HideSolutionNode = FALSE
@@ -90,14 +90,14 @@ Global
 	GlobalSection(NestedProjects) = preSolution
 		{D32BA840-A58F-441F-9211-2B50646434B9} = {BDD76646-2418-41BA-8F7A-4CB3C2B42593}
 		{AE939C31-706D-47C2-8445-9A5141AD6A37} = {BDD76646-2418-41BA-8F7A-4CB3C2B42593}
-		{64B7C964-21DA-4612-9317-35E0F5562BE1} = {3E03BA6A-76DC-4B42-8C54-99FDD716ED70}
-		{04EFC854-905D-4875-85EB-0AAEEFFE5F47} = {3E03BA6A-76DC-4B42-8C54-99FDD716ED70}
 		{9CF4DB78-78F3-4E40-A60B-807BE00CD02C} = {BDD76646-2418-41BA-8F7A-4CB3C2B42593}
 		{61CCF9C3-BA54-46D1-883A-BDBD5B3508DB} = {BDD76646-2418-41BA-8F7A-4CB3C2B42593}
 		{49E61955-B655-4752-ABD1-D7802B69BF44} = {16FA0B29-C6FF-46A8-BB2D-AC9CA744BA07}
 		{F9B10063-5389-4F5E-8F2B-4C2E46988EC6} = {E437EDF2-E0D2-4CB7-885E-E500019E0AFD}
 		{E1DE7318-1459-4AAD-9A36-90DE438BD4C6} = {E437EDF2-E0D2-4CB7-885E-E500019E0AFD}
 		{FEC0BB53-CB6D-48F2-BBF8-6D82AF5B7647} = {E437EDF2-E0D2-4CB7-885E-E500019E0AFD}
+		{1E64FEBC-B19E-4C94-9F5B-39CB449D0FB7} = {3E03BA6A-76DC-4B42-8C54-99FDD716ED70}
+		{AE5EEFC7-B685-41AF-89D0-C2DDB925C507} = {3E03BA6A-76DC-4B42-8C54-99FDD716ED70}
 	EndGlobalSection
 	GlobalSection(ExtensibilityGlobals) = postSolution
 		SolutionGuid = {35AE73FF-F868-4905-9C15-F260BAE28227}
